Implement SharedArrayBuffer for all typed arrays
# Conflicts:
#	tests/testsrc/test262.properties

+++ rhino/src/main/java/org/mozilla/javascript/typedarrays/NativeBigInt64Array.java
@@ -54,7 +54,7 @@ public class NativeBigInt64Array extends NativeBigIntArrayView {
     }
 
     public NativeBigInt64Array(int len) {
-        this(new NativeArrayBuffer((double) len * BYTES_PER_ELEMENT), 0, len);
+        this(new NativeArrayBuffer(len * BYTES_PER_ELEMENT), 0, len);
     }
 
     @Override
